study-app-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg[binary]==3.1.14
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
google-auth==2.25.2
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pdfplumber==0.10.3
python-multipart==0.0.6
aiosqlite==0.19.0
